requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.10
pydantic-settings==2.5.2
email-validator==2.2.0
qdrant-client==1.11.3
langchain-text-splitters==0.3.0
sentence-transformers==3.1.1
pypdf==5.0.1
python-docx==1.1.2
